backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.5",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asyncmy==0.2.10",
    "bcrypt==4.3.0",
    "cffi==1.17.1",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.7",
    "dnspython==2.7.0",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "mysql-connector-python==9.0.0",
    "passlib==1.7.4",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pyjwt==2.4.0",
    "pymysql==1.1.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.32",
    "starlette==0.46.2",
    "tomli==2.2.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "uvicorn>=0.35.0",
]
